DAS-8: DAS-21 & DAS-22

+++ src/main/java/dk/northtech/dasscofileproxy/webapi/v1/FtpsClient.java
@@ -1,21 +1,29 @@
 package dk.northtech.dasscofileproxy.webapi.v1;
 
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.interfaces.DecodedJWT;
 import dk.northtech.dasscofileproxy.domain.Asset;
+import dk.northtech.dasscofileproxy.domain.AssetFull;
+import dk.northtech.dasscofileproxy.service.AssetService;
 import dk.northtech.dasscofileproxy.service.FtpsService;
+import dk.northtech.dasscofileproxy.service.KeycloakService;
+import jakarta.annotation.security.RolesAllowed;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
-import jakarta.ws.rs.container.AsyncResponse;
-import jakarta.ws.rs.container.Suspended;
-import jakarta.ws.rs.core.MediaType;
-import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.*;
 import org.apache.commons.net.ftp.FTPFile;
+import org.json.JSONArray;
+import org.json.JSONObject;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
+import org.springframework.web.client.RestTemplate;
 
 import java.io.*;
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
-import java.util.function.Supplier;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 @Component
 @Path("/v1/ftps")
@@ -23,9 +31,12 @@ public class FtpsClient {
 
     private final FtpsService ftpsService;
 
+    private final AssetService assetService;
+
     @Inject
-    public FtpsClient(FtpsService ftpsService) {
+    public FtpsClient(FtpsService ftpsService, AssetService assetService) {
         this.ftpsService = ftpsService;
+        this.assetService = assetService;
     }
 
 
@@ -38,23 +49,76 @@ public class FtpsClient {
         this.ftpsService.close();
         return files;
     }
+
     @GET
-    @Path("{path : .+}")
+    @Path("{params : .+}")
     @Produces(MediaType.APPLICATION_OCTET_STREAM)
-    public Response get(@Encoded @PathParam("path") String path) throws IOException {
+    public Response get(@Encoded @PathParam("params") String params, @Context HttpHeaders headers) throws IOException {
+
+        // Split path param into guid and filename
+        var strings = params.split("/");
+        // Throw exception if path param doesn't contain exactly two arguments
+        if (strings.length != 2)
+            return Response.status(Response.Status.BAD_REQUEST).entity("Path must consist of a asset guid and filename using the following format \"{assetGUID}/{filename.extension}\"").build();
+        var guid = strings[0];
+        var file = strings[1];
+
+        // Read the asset from the asset service to obtain restricted status and remaining path for the asset
+        AssetFull assetFull = assetService.getFullAsset(guid);
+
+        // If there is no asset for the guid in asset service throw exception, as the client needs institution and collection to locate file
+        if (assetFull == null)
+            return Response.status(Response.Status.NOT_FOUND).entity("Cannot find asset in asset service.").build();
+
+        // Determine whether the user should be allowed access to the file
+        AtomicBoolean userHasAccess = new AtomicBoolean(false);
+        if (!assetFull.restricted_access.isEmpty()) {
+            if (headers == null || headers.getHeaderString(HttpHeaders.AUTHORIZATION) == null) return Response.status(Response.Status.UNAUTHORIZED).entity("The requested asset has restricted access and the request does not include valid authorization header.").build();
+
+            String authHeader = headers.getHeaderString(HttpHeaders.AUTHORIZATION);
+            String[] parts = authHeader.split(" ");
+            if (parts.length!=2) return Response.status(Response.Status.BAD_REQUEST).entity("Authorization header must comply with basic auth headers, i.e. the headers content should look like \"Bearer {token}\".").build();
+            DecodedJWT jwt = JWT.decode(parts[1]);
+
+            System.out.println(jwt.getClaims().get("realm_access").toString());
+            String rolesStr = jwt.getClaims().get("realm_access").toString();
+            String toConvert = rolesStr.substring(rolesStr.indexOf("["));
+            toConvert = toConvert.substring(0, toConvert.indexOf("]")+1);
+            JSONArray rolesArray = new JSONArray(toConvert);
+
+            // Convert the JSONArray to a List of Strings
+            List<String> roles = new ArrayList<>();
+            for (int i = 0; i < rolesArray.length(); i++) {
+                roles.add(rolesArray.getString(i));
+            }
+
+            assetFull.restricted_access.forEach(role -> {
+                if (roles.contains(role.name())) userHasAccess.set(true);
+            });
+
+
+        } else userHasAccess.set(true);
+        if (!userHasAccess.get()) return Response.status(Response.Status.UNAUTHORIZED).entity("The requested asset has restricted access and the user does not have any of the required roles.").build();
+
+        // Build full path for requested file
+        String path = "DaSSCoStorage/" + assetFull.institution + "/" + assetFull.collection + "/" + guid + "/" + file;
+
         // Check if the file is already cached
+        // TODO: Fix caching after demo
+        /*
         InputStream cached = ftpsService.getCached("cached/" + path);
         if (cached != null) {
             // Refresh the time to live for the cached file
             ftpsService.refreshTimeToLive(path);
             return Response.ok(cached).build(); // Return the cached file
         }
+         */
 
         // Open the FTP service connection
         ftpsService.open();
 
         // Download the file from the FTP server
-        InputStream fileStream = ftpsService.downloadFile("DaSSCoStorage/" + path);
+        InputStream fileStream = ftpsService.downloadFile(path);
         if (fileStream == null) {
             // File not found on the FTP server
             return Response.status(404).entity("The requested resource could not be found.").build();
@@ -62,7 +126,7 @@ public class FtpsClient {
 
         // Asynchronously start caching file
         CompletableFuture.supplyAsync(() -> {
-            ftpsService.cacheFile("DaSSCoStorage/" + path);
+            ftpsService.cacheFile(path);
             return null;
         });
 
@@ -70,11 +134,10 @@ public class FtpsClient {
         return Response.ok(fileStream).build();
     }
 
-
     @PUT
     @Path("upload/{path}")
     @Produces(MediaType.APPLICATION_JSON)
-    public boolean update(@PathParam("path")String path, Asset asset) {
+    public boolean update(@PathParam("path") String path, Asset asset, @QueryParam("cleanup") boolean cleanup) {
         try {
             // Open the FTPS service connection
             this.ftpsService.open();
@@ -131,13 +194,23 @@ public class FtpsClient {
                 }
 
                 // Delete the local files
-                for (File file : files) {
-                    if (!file.isDirectory()) {
-                        var deleted = file.delete();
-                        if (!deleted) {
-                            throw new RuntimeException("Couldn't delete file: " + file.getName() + ".");
-                        }
+                boolean containsDirectory = false;
+                if (cleanup) {
+                    for (File file : files) {
+                        if (!file.isDirectory()) {
+                            var deleted = file.delete();
+                            if (!deleted) {
+                                throw new RuntimeException("Couldn't delete file: " + file.getName() + ".");
+                            }
+                        } else containsDirectory = true;
                     }
+
+                    // Don't attempt to delete folder if it contains other folders
+                    if (!containsDirectory) {
+                        var deleted = folder.delete();
+                        if (!deleted) throw new RuntimeException("Couldn't delete folder: " + folder.getName() + ".");
+                    }
+
                 }
             }
